requirements.txt
streamlit
streamlit-folium
pandas
numpy
Pillow
pyproj
haversine
Shapely
tqdm
geopandas
geojson
protobuf
requests
requests-html
pyodbc
